ai/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
python-dotenv==1.0.1
langchain==0.3.7
langchain_community==0.3.7
langchain_core>=0.3.27
langchain_huggingface==0.1.2
langchain_google_genai==2.0.8
jq==1.8.0